+++ R/cdr.R
@@ -85,6 +85,13 @@ cdr_variables_daily <- list(
     stdev = "cdr_seaice_conc_stdev",
     interpolation_spatial = "cdr_seaice_conc_interp_spatial_flag",
     interpolation_temporal = "cdr_seaice_conc_interp_temporal_flag"
+  ),
+  v6 = list(
+    aice = "cdr_seaice_conc",
+    qa = "cdr_seaice_conc_qa_flag",
+    stdev = "cdr_seaice_conc_stdev",
+    interpolation_spatial = "cdr_seaice_conc_interp_spatial_flag",
+    interpolation_temporal = "cdr_seaice_conc_interp_temporal_flag"
   )
 )
 
@@ -101,8 +108,16 @@ cdr_variables_monthly <- list(
     aice = "cdr_seaice_conc_monthly",
     stdev = "cdr_seaice_conc_monthly_stdev",
     qa = "cdr_seaice_conc_monthly_qa_flag"
+  ),
+
+  v6 = list(
+    aice = "cdr_seaice_conc_monthly",
+    stdev = "cdr_seaice_conc_monthly_stdev",
+    qa = "cdr_seaice_conc_monthly_qa_flag"
   )
 )
+
+supported_versions <- c(4, 5, 6)
 
 cdr_variables <- list(
   monthly = cdr_variables_monthly,
@@ -210,8 +225,10 @@ cdr <- function(
   dir = tempdir(),
   use_cache = TRUE
 ) {
-  if (!(version %in% c(4, 5))) {
-    cli::cli_abort("Invalid version. Available versions are 4 and 5.")
+  if (!(version %in% supported_versions)) {
+    cli::cli_abort(
+      "Invalid version. Available versions are {supported_versions}"
+    )
   }
 
   resolution <- resolution[1]