+++ NEWS.md
@@ -1,3 +1,5 @@
+# icecdr (development version)
+
 # icecdr 1.0.0
 
 # icecdr 0.1.0