+++ NEWS.md
@@ -1,5 +1,9 @@
 # icecdr (development version)
 
+## New features
+
+-   The new `sea_ice_index()` function downloads monthly and daily sea ice extent.
+
 # icecdr 1.0.0
 
-* Initial CRAN submission.
+-   Initial CRAN submission.