+++ NEWS.md
@@ -1,3 +1,5 @@
+# icecdr (development version)
+
 # icecdr 1.1.0
 
 ## New features