+++ NEWS.md
@@ -1,6 +1,8 @@
 # icecdr (development version)
 
-Switches to the CoastWatch ERDDAP due to PolarWatch closure.
+-   Switches to the CoastWatch ERDDAP due to PolarWatch closure.
+
+-   Adds version 6.
 
 # icecdr 1.1.0