+++ NEWS.md
@@ -1,4 +1,4 @@
-# icecdr (development version)
+# icecdr 1.2.0
 
 -   Switches to the CoastWatch ERDDAP due to PolarWatch closure.